requirements.txt
opencv-python
numpy
matplotlib
ipywidgets
Pillow
torch
torchvision
tqdm
fastapi
uvicorn
python-multipart